+++ rust/src/p0005.rs
@@ -15,9 +15,9 @@ numbers from 1 to 20?
 */
 use itertools::Itertools;
 
-pub fn p0005() -> u64 {
-    let mut answer = u64::MAX;
-    let group = 1..=20u64;
+pub fn p0005() -> i128 {
+    let mut answer = u32::MAX;
+    let group = 1..=20u32;
     for x in group.clone() {
         for multiples in group.clone().combinations(x as usize) {
             let num = multiples.into_iter().product();
