+++ rust/src/p0008.rs
@@ -35,7 +35,7 @@ greatest product. What is the value of this product?
 */
 
 
-pub fn p0008() -> u64 {
+pub fn p0008() -> i128 {
     let string = concat!(
         "73167176531330624919225119674426574742355349194934",
         "96983520312774506326239578318016984801869478851843",
@@ -58,12 +58,12 @@ pub fn p0008() -> u64 {
         "05886116467109405077541002256983155200055935729725",
         "71636269561882670428252483600823257530420752963450"
     );
-    let mut answer: u64 = 0;
+    let mut answer: u32 = 0;
     for i in 0..(string.len() - 13) {
         let slice = &string[i..(i + 13)];
-        let mut prod: u64 = 1;
+        let mut prod: u32 = 1;
         for c in slice.bytes() {
-            prod *= (c as u32 - '0' as u32) as u64;
+            prod *= c as u32 - '0' as u32;
         }
         if prod > answer {
             answer = prod;
